Read screen has time interval
Updated the frontend to display the last time the backend got new data and the elapsed time.

This required a change to the mySQL backend, the API to write, the API to read.  Use mach_num to be 99 in the mySQL.  API read throws in another mach_num as 98 to be the elapsed time.  This means the front end does not need moment.js or other special loads.